requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-whois==0.8.0
scapy==2.5.0
colorama==0.4.6 